feat: Add integration tests for account queries to validate retrieval by Id and UserId
